Relocate eks related code to internal package
