+++ internal/cluster/clusteradapter/node_pool_processor.go
@@ -22,7 +22,7 @@ import (
 	"github.com/mitchellh/mapstructure"
 
 	"github.com/banzaicloud/pipeline/internal/cluster"
-	"github.com/banzaicloud/pipeline/internal/cluster/distribution"
+	eks2 "github.com/banzaicloud/pipeline/internal/cluster/distribution/eks"
 	"github.com/banzaicloud/pipeline/internal/providers/amazon/amazonadapter"
 	"github.com/banzaicloud/pipeline/pkg/cluster/eks"
 	"github.com/banzaicloud/pipeline/pkg/providers"
@@ -48,7 +48,7 @@ func (v DistributionNodePoolProcessor) ProcessNew(
 ) (cluster.NewRawNodePool, error) {
 	switch {
 	case c.Cloud == providers.Amazon && c.Distribution == "eks":
-		var nodePool distribution.NewEKSNodePool
+		var nodePool eks2.NewNodePool
 
 		err := mapstructure.Decode(rawNodePool, &nodePool)
 		if err != nil {
